Refactor: Restructure Azure components into organized subdirectory
## Overview
Refactored the large AzureContainerList component (1,059 lines) into a well-organized, maintainable component structure following the pattern established by other feature domains (postgres, deployments).

## Changes Made

### Directory Structure
- Created `client/src/components/azure/` subdirectory
- Organized components by responsibility with clear separation of concerns

### Component Breakdown
1. **Main Components** (moved to azure/)
   - `AzureContainerList.tsx` - Refactored from 1,059 to 285 lines
   - `AzureConnectivityStatus.tsx` - Moved without changes

2. **Sub-Components** (new)
   - `AzureContainerFilters.tsx` - Filter controls (95 lines)
   - `AzureContainerPagination.tsx` - Pagination UI (137 lines)
   - `AzureContainerTable.tsx` - Table rendering (120 lines)

3. **Cell Components** (azure/cells/)
   - `CopyButton.tsx` - Reusable copy button (35 lines)
   - `ContainerNameCell.tsx` - Container name display (24 lines)
   - `LastModifiedCell.tsx` - Date formatting cell (33 lines)
   - `LeaseStatusCell.tsx` - Lease status badge (26 lines)
   - `PublicAccessCell.tsx` - Access level badge (26 lines)
   - `MetadataCell.tsx` - Metadata indicator (46 lines)
   - `ActionsCell.tsx` - Action buttons with test status (72 lines)
   - `index.ts` - Barrel export for cells

4. **Utilities & Hooks**
   - `constants.ts` - Badge configs, types, column widths (90 lines)
   - `hooks/use-container-columns.tsx` - Column definitions hook (106 lines)
   - `utils/container-filters.ts` - Filter & pagination logic (112 lines)

5. **Barrel Export**
   - `index.ts` - Centralized exports for clean imports

### Benefits
- **Maintainability**: Reduced main component from 1,059 to 285 lines (~73% reduction)
- **Reusability**: Extracted 7 cell components and shared utilities
- **Testability**: Smaller, focused components are easier to test
- **Consistency**: Follows existing patterns (postgres/, deployments/ structure)
- **Developer Experience**: Better file organization and imports via barrel exports

### Import Changes
- Updated `client/src/app/connectivity/azure/page.tsx` to use new import path
- Changed from `@/components/AzureContainerList` to `@/components/azure`

### Build Verification
- ✅ All TypeScript compilation successful
- ✅ Build completed without errors
- ✅ No breaking changes to public API
