+++ client/src/components/azure/AzureConnectivityStatus.tsx
